ws-img/requirements.txt
pymongo==3.9.0
Pillow==6.0.0
Flask==1.1.1
send_files==0.1